Fix comparison warning

+++ src/comm.cpp
@@ -8,7 +8,7 @@ Packet readPacket()
     packet.Data = (uint8_t*)malloc(packet.Size);
     if(packet.Data != NULL)
     {
-        int bytesRead = 0;
+        size_t bytesRead = 0;
         while(bytesRead < packet.Size)
         {
             bytesRead += Serial.readBytes(packet.Data + bytesRead, packet.Size - bytesRead);
